requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cycler==0.11.0
decorator==4.4.2
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
fonttools==4.28.5
idna==3.3
imageio==2.14.1
imageio-ffmpeg==0.4.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
moviepy==1.0.3
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
proglog==0.1.9
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
tqdm==4.62.3
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.2
